Reorganized a few files
- Moved the caches to a separate file.
- Moved the message definitions to a separate file.
