+++ anydex/core/defs.py
@@ -1,0 +1,15 @@
+# Message definitions
+MSG_MATCH = 7
+MSG_MATCH_DECLINE = 9
+MSG_PROPOSED_TRADE = 10
+MSG_DECLINED_TRADE = 11
+MSG_COUNTER_TRADE = 12
+MSG_ACCEPT_TRADE = 13
+MSG_ORDER_QUERY = 16
+MSG_ORDER_RESPONSE = 17
+MSG_BOOK_SYNC = 19
+MSG_PING = 20
+MSG_PONG = 21
+MSG_MATCH_DONE = 22
+MSG_PK_QUERY = 23
+MSG_PK_RESPONSE = 24
